spike(studio): evaluate @sentry/tanstackstart-react unified SDK
Replaces Studio's hand-rolled Sentry wiring on the TanStack Start build with
the official unified `@sentry/tanstackstart-react` SDK (pinned to 10.59.0 to
match the existing @sentry/nextjs / @sentry/react versions):

- compat/sentry-nextjs.ts + vite alias: re-point the app's ubiquitous
  `@sentry/nextjs` imports to the unified SDK in the TanStack build (single
  SDK / single @sentry/core hub). Drops the @sentry/nextjs SSR-bundling
  workarounds since the unified SDK's ESM server entry exposes real named
  exports.
- lib/sentry-client-config.ts: shared client init factory + noise predicates,
  extracted from instrumentation-client.ts so both runtimes gate identically.
- router.tsx: client Sentry.init (client-only) via the shared factory + router
  browser tracing.
- instrument.server.mjs + start.ts + api/server.js + scripts/serve.js: server
  init, global request/function middleware, and wrapFetchWithSentry.
- vite.config.ts: sentryTanstackStart() plugin appended last.

TanStack build + server-boot smoke pass; tsc clean; client tests pass. Next.js
build path (instrumentation.ts / sentry.server.config.ts / next.config.ts)
left intact. See PR body for the difficulty assessment + friction points.
